Make event_handlers into a dict of dicts
When we added the subscription id to allow for unsubscription we left
this data structure as a dict of sets.  However, there are really two
key values.  The first key remains the event.  The second key is the
subscription id.  A dict of dicts allows us to represent both of these
smoothly instead of having to iterate through a set of (sub_id, handler)
tuples.

This change also fixes a problem where events were not always removed
when a callback was deallocated.  (Callbacks can be deallocated outside
of our code because we hold a weakref to them so as not to leak memory)